quiz page created to progress to L03
L02 is not fully done yet, but I created a demo of what the quiz page will be. that code should not change. I added a room feature with a quiz code, and intend to allows users that have joined that room to created questions for the quiz until the game has been started. still working on that feature and the timer, but this should be enough to work with hopefully. the see the room feature use /obj2 or /room to see the quiz feature use /quiz. any other questions ill answer in the discord.