obj2 backend added, needs to be merged and pushed to main
start game needs to redirect to the quiz , copy and paste ur code to display quiz questions in room. also smthn with the post broke in the mean time. i think we should revert to their version (or yours if all other features are working) and then copy and paste my work in. users grades are displayed, manually scored using js and stored in the db.